% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheConnection.R
\docType{class}
\name{CacheConnection}
\alias{CacheConnection}
\title{CacheConnection Class}
\format{
An [R6::R6Class] generator object.
}
\description{
An R6 class that handles persistent or in-memory caching of data used in the Countdown 2030
analysis and reporting workflows. It supports tracking of various internal data objects,
reactive updates (for Shiny apps).
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{county}}{description}

\item{\code{year_type}}{description}

\item{\code{year}}{description}

\item{\code{aggregation_unit}}{description}

\item{\code{aggregation_level}}{description}

\item{\code{k_factors}}{description}

\item{\code{khis_data}}{description}

\item{\code{summarised_data}}{description
Update a field (with change tracking)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheConnection-new}{\code{CacheConnection$new()}}
\item \href{#method-CacheConnection-reactive}{\code{CacheConnection$reactive()}}
\item \href{#method-CacheConnection-set_county}{\code{CacheConnection$set_county()}}
\item \href{#method-CacheConnection-set_year_type}{\code{CacheConnection$set_year_type()}}
\item \href{#method-CacheConnection-set_year}{\code{CacheConnection$set_year()}}
\item \href{#method-CacheConnection-set_aggregation_unit}{\code{CacheConnection$set_aggregation_unit()}}
\item \href{#method-CacheConnection-set_aggregation_level}{\code{CacheConnection$set_aggregation_level()}}
\item \href{#method-CacheConnection-clone}{\code{CacheConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CacheConnection instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-reactive"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-reactive}{}}}
\subsection{Method \code{reactive()}}{
Return a reactive wrapper (for Shiny).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$reactive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_county"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_county}{}}}
\subsection{Method \code{set_county()}}{
Set county.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_county(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_year_type"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_year_type}{}}}
\subsection{Method \code{set_year_type()}}{
Set county.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_year_type(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_year"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_year}{}}}
\subsection{Method \code{set_year()}}{
Set county.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_year(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_aggregation_unit"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_aggregation_unit}{}}}
\subsection{Method \code{set_aggregation_unit()}}{
Set county.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_aggregation_unit(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_aggregation_level"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_aggregation_level}{}}}
\subsection{Method \code{set_aggregation_level()}}{
Set county.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_aggregation_level(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
