% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheConnection.R
\name{init_CacheConnection}
\alias{init_CacheConnection}
\title{Create a CacheConnection Object}
\usage{
init_CacheConnection()
}
\value{
An instance of the `CacheConnection` class.
}
\description{
`init_CacheConnection` initializes a `CacheConnection`.
}
