% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moh.analytics-package.R
\docType{package}
\name{moh.analytics-package}
\alias{moh.analytics}
\alias{moh.analytics-package}
\title{moh.analytics: Show Quick Analytics for Selected Indicators}
\description{
An internal Shiny application to quickly visualize selected health indicators. Built for use within the Ministry of Health to support timely decision-making and performance tracking.
}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com}

}
\keyword{internal}
